server/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
pymongo==4.10.1
python-dotenv==1.0.1